+++ synthesizer/__init__.py
@@ -1,2 +1,3 @@
 from .player import Player
 from .synthesizer import Oscillator, Synthesizer, Waveform
+from .writer import Writer